+++ src/components/index.ts
@@ -89,6 +89,7 @@ export { default as FlowRunsScatterPlot } from './FlowRunsScatterPlot.vue'
 export { default as FlowRunsSort } from './FlowRunsSort.vue'
 export { default as FlowRunStartTime } from './FlowRunStartTime.vue'
 export { default as FlowRunSubFlows } from './FlowRunSubFlows.vue'
+export { default as FlowRunTagCombobox } from './FlowRunTagCombobox.vue'
 export { default as FlowRunTaskCount } from './FlowRunTaskCount.vue'
 export { default as FlowRunTaskRuns } from './FlowRunTaskRuns.vue'
 export { default as FlowRunTimeline } from './FlowRunTimeline.vue'
@@ -185,7 +186,6 @@ export { default as StateListItem } from './StateListItem.vue'
 export { default as StateNameSelect } from './StateNameSelect.vue'
 export { default as StateSelect } from './StateSelect.vue'
 export { default as SubmitButton } from './SubmitButton.vue'
-export { default as TagCombobox } from './TagCombobox.vue'
 export { default as TaskRunDetails } from './TaskRunDetails.vue'
 export { default as TaskRunList } from './TaskRunList.vue'
 export { default as TaskRunListItem } from './TaskRunListItem.vue'
